+++ src/Phobos.h
@@ -34,6 +34,7 @@ public:
 	static bool IsLoadingSaveGame;
 	static bool ShouldSave;
 	static std::wstring CustomGameSaveDescription;
+	static void ScheduleGameSave(const std::wstring& description);
 	static void PassiveSaveGame();
 #ifdef DEBUG
 	static bool DetachFromDebugger();